requirement.txt
streamlit
pandas
python-dotenv
huggingface_hub
scikit-learn